Add more type-casts, code cleanup (all backported from 8.7)

+++ macosx/tkMacOSXClipboard.c
@@ -3,9 +3,9 @@
  *
  *	This file manages the clipboard for the Tk toolkit.
  *
- * Copyright (c) 1995-1997 Sun Microsystems, Inc.
- * Copyright 2001-2009, Apple Inc.
- * Copyright (c) 2006-2009 Daniel A. Steffen <das@users.sourceforge.net>
+ * Copyright © 1995-1997 Sun Microsystems, Inc.
+ * Copyright © 2001-2009 Apple Inc.
+ * Copyright © 2006-2009 Daniel A. Steffen <das@users.sourceforge.net>
  *
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -37,7 +37,7 @@ static Tk_Window clipboardOwner = NULL;
 			cbPtr; cbPtr = cbPtr->nextPtr) {
 		    NSString *s = [[TKNSString alloc]
 				      initWithTclUtfBytes:cbPtr->buffer
-						   length:cbPtr->length];
+						   length:(NSUInteger)cbPtr->length];
 		    [string appendString:s];
 		    [s release];
 		}
